// src/app/dashboard/import-page/file-upl/models/file-upl.model.ts
export class FileUpl {
  dateListAsInt: string = '';
  name: string = '';
  surname: string = '';
  linkedin: string = '';
  comment: string = '';
  technologyListAsString: string = '';
  whenToContact: string = '';
  open: string = '';
}

export class TestUpl {
  email: string = '';
  password: string = '';
}